got the home routes functional and takin in data
