+++ package.json
@@ -13,12 +13,13 @@
     "connect-session-sequelize": "^7.1.2",
     "dotenv": "^16.0.0",
     "eslint": "^8.9.0",
-    "express": "^4.17.2",
+    "express": "^4.17.3",
     "express-handlebars": "^6.0.2",
     "express-session": "^1.17.2",
+    "handlebars": "^4.7.3",
     "jest": "^27.5.1",
     "mysql2": "^2.3.3",
-    "sequelize": "^6.16.1",
+    "sequelize": "^6.16.2",
     "yarn": "^1.22.17"
   },
   "scripts": {
