+++ views/coding-challenges.handlebars
@@ -6,9 +6,9 @@
     <section>
         <h2>CHALLENGES</h2>
         <ol>
-            {{#each codes as |code|}}
+            {{#each challenges as |challenge|}}
             <li>
-                {{> code-info code}}
+                {{> code-info challenge}}
             </li>
             {{/each}}
         </ol>
