+++ views/login.handlebars
@@ -31,5 +31,3 @@
 </form>
 
 <script src="/javascript/login.js"></script>
-
-<script src="/javascript/login.js"></script>
