+++ views/partials/code-info.handlebars
@@ -1,8 +1,9 @@
 <div class="code">
   <a href="{{code_url}}" target="_blank">{{title}}</a>
+  <p>{{title}}</p>
   <p>Submitted by:
     {{user.username}}
     | on:
-    {{created_at}}
+    {{createdAt}}
   </p>
 </div>
